Fake Furnace, Texturen Fehlen und nur Interface wird angezeigt

+++ src/main/java/net/chris/bastimod/block/custom/FakeFurnaceBlock.java
@@ -53,7 +53,7 @@ public class FakeFurnaceBlock extends AbstractFurnaceBlock {
     public void animateTick(BlockState pState, Level pLevel, BlockPos pPos, RandomSource pRandom) {
         if (pState.getValue(LIT)) {
             double d0 = (double)pPos.getX() + 0.5D;
-            double d1 = (double)pPos.getY();
+            double d1 = pPos.getY();
             double d2 = (double)pPos.getZ() + 0.5D;
             if (pRandom.nextDouble() < 0.1D) {
                 pLevel.playLocalSound(d0, d1, d2, SoundEvents.FURNACE_FIRE_CRACKLE, SoundSource.BLOCKS, 1.0F, 1.0F, false);
